Only author of question can removes the link

+++ app/controllers/links_controller.rb
@@ -1,7 +1,9 @@
 class LinksController < ApplicationController
+  before_action :authenticate_user!
+
   expose :link
 
   def destroy
-    link.destroy
+    link.destroy if current_user.author_of?(link.linkable)
   end
 end
